requirements.txt
flask
pillow 
qrcode[pil]
requests
replicate
python-dotenv